FreeRTOS-Plus-TCP: Port a new qemu virtio-net interface
The port is derived from the porting guide [1] with modifications to work for
QEMU's legacy virtio-net. It has been tested on qemu-system-riscv64 with
UDP/TCP echo, http servers and clients via a tap interface.

It's the inefficient non-zero copy example in [1]. It can work in both
interrupt-based and polling-based modes, though polling-based has only been
tested so far.

[1] https://www.freertos.org/FreeRTOS-Plus/FreeRTOS_Plus_TCP/Embedded_Ethernet_Porting.html